+++ globals.d.ts
@@ -1,5 +1,0 @@
-declare global {
-    interface Window {
-        SP_REACT: typeof React;
-    }
-}